pyproject.toml
[tool.poetry]
package-mode = true
name = "fastapi-assignment"
version = "0.1.0"
description = ""
authors = ["Vineet Sharma <vin8003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.20"
fastapi = "^0.115.4"
pydantic = {extras = ["email"], version = "^2.9.2"}
motor = "^3.6.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest-asyncio = "^0.24.0"
redis = "^4.0.0"
sentry-sdk = "^2.17.0"
uvicorn = "^0.32.0"
fastapi-mail = "^1.4.1"
celery = {extras = ["async"], version = "^5.4.0"}
httpx = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
pytest-mock = "^3.14.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
